requirements.txt
python-dotenv>=1.0.0
pytz>=2025.2
asyncio==4.0.0
requests==2.32.5
tenacity>=9.1.2

# Forked edgeX official python sdk with post_only limit order
git+https://github.com/your-quantguy/edgex-python-sdk.git@7ce85b6ebfa9df0a6f6252d93e10d106927257a9#egg=edgex-python-sdk

# Lighter exchange SDK
git+https://github.com/elliottech/lighter-python.git@d000979970aad54ebb940aa3dc90cbc00028c54

# Backpack Exchange SDK
bpx-py==0.2.0